Update testing for road trip

+++ spec/requests/api/v1/forecast_spec.rb
@@ -30,7 +30,7 @@ describe "forecast return happy path" do
       expect(body[:data][:attributes][:current_weather][:temperature]).to be_a(Float)
       expect(body[:data][:attributes][:current_weather][:feels_like]).to be_a(Float)
       expect(body[:data][:attributes][:current_weather][:humidity]).to be_a(Integer)
-      expect(body[:data][:attributes][:current_weather][:uvi]).to be_a(Integer)
+      expect(body[:data][:attributes][:current_weather][:uvi]).to be_a(Float)
       expect(body[:data][:attributes][:current_weather][:visibility]).to be_a(Integer)
       expect(body[:data][:attributes][:current_weather][:conditions]).to be_a(String)
       expect(body[:data][:attributes][:current_weather][:icon]).to be_a(String)
